+++ TiendaApiDto/Repositories/IProductoRepository.cs
@@ -9,7 +9,7 @@ namespace TiendaApiDto.Repositories
         Task<IEnumerable<Producto>> GetAllAsync();
         Task<Producto?> GetByIdAsync(int id);
         Task<Producto> AddAsync(Producto producto);
-        Task<Producto?> UpdateAsync(long id, Producto producto);
-        Task<bool> DeleteAsync(long id);
+        Task<Producto?> UpdateAsync(int id, Producto producto);
+        Task<bool> DeleteAsync(int id);
     }
 }
